cntools v2
Third iteration of the cntools rewrite.
As its a complete rewrite but based on the core of cntools by gufmar reviewing all changes would be a nightmare, so I instead created a new script called cntools-v2.sh.
It is now more interactive instead of requiring user to input all arguments on command line. This should make it less prone for users to input wrong argument at wrong position.