[CNTOOLS] Block Collector now also parsing for non adopted blocks
Parser updated to look for TraceNodeIsLeader and TraceForgedInvalidBlock in addition to previous TraceAdoptedBlock. If TraceForgedInvalidBlock is found the entire json trace is base64 encoded and logged.

Non adopted blocks that are not adopted but also not invalid(fork?) is displayed with null value for size and hash.

A timeout of 4s is added to ledger-state dump in cntools to not hang forever.