CNTools wallet redesign [breaking change]
Support for payment address(aka enterprise) removed and only base address support left. Everything that can be done with a payment address can also be done with the base address, ie doesn't need to be registered on chain unless its delegated or just as pledge wallet. This is done to present fewer options to end user and remove confusion regarding the different address types. Because of this, commit is set as a breaking change.

Previous payment wallets(ie not upgraded to stake wallet) wont be listed in CNTools. Either send all funds to a stake wallet or use external script to move funds afterwards.

Wallet registration is removed as a manual action and is now done automatically when delegating the wallet or using it as pledge wallet when creating/modifying a pool. This also goes for genesis address, use external script to send funds to a CNTools created wallet.