Delete helper function created to reduce duplicate code.
Even though secure delete has limited effect on modern hardware and file systems as mentioned in PR, adding support for srm shouldn't have any negative effects. The check is also made silent, if srm is not available it will continue to use rm without any warning.

CNTools wont be able to secure the keys completely, though built in encrypt option and/or backup with option to remove keys from disk adds a good first step in securing the system. It's up to the end user to improve on this and implement good security measures by utilizing full disk encryption and/or other ways to securely handle the keys and physical access to server.