add cncli installation and run script (#572)
CNCLI is developed by Andrew and available at https://github.com/AndrewWestberg/cncli
It's a RUST implementation for low-level interaction with cardano-node.
Current functionality includes cardano-node ping and chain sync into sqlite db for different block queries but is planned to be extended with more features.

Use cncli-install.sh to install/update RUST, CNCLI, and deploy it as a systemd service. Updated prereqs.sh needed to install libsqlite3 and env updated with a variable for path to CNCLI executable.

gLiveView updated to v1.10 to utilize CNCLI ping
For outgoing, if CNCLI is missing, fall back on tcptraceroute. If that's also missing check with ICMP ping as a last result.
A flag is added [`-p`] to be able to disable the default CNCLI ping.