Move all storage except cardano-node db out of /db into /guild-db
It was observed that in case of a cardano-node db wipe and sync from zero the node require an empty db folder to work. Because of this all guilds extra storage previously placed in /db folder is now put in /guild-db instead.

This affect shelley_trans_epoch file, logMonitor(cntoolsBlockCollector) block folder and cncli sqlite db.