fix for sqlite3 -column bug
found a bug in sqlite3 output when -column was used. value was truncated to the size(number of characters) of the value in the first row. -column was used in scripts to query data and split each column into bash variables. Now replaced with IFS='|' in while loops to split on | instead of whitespace.