cncli leaderlog fix
Previously ledger-state was only dumped once per epoch in cncli leaderlog to save on processing. If this was done early in the epoch the proposal for next epoch d decrement might not have been available and thus the calculations are incorrect. This change will dump a new ledger-state file if the previous one is older than 60min or non-existing.