CNCLI Sendslots (#610)
Add support for PoolTool sendslots implemented in CNCLI 0.3.1. A new systemd service file added to deploy script.

Changes added is that it now gives the user the option to deploy logmonitor or not. Compared to previously when it installed it but never activated the service.

PoolTool sendslots is also a user option when deploying Blocklog. Blocklog services is required for sendslots to work(sync + leaderlog).