cncli.sh ptsendslots epoch boundary improvement (#777)
cncli.sh ptsendslots tried to send slots as soon as it saw a new epoch but due to node load at boundary this could fail. The failure was not detected properly and slots was not sent in this case. CNCLI report error with json output, not exit code.

Now waits 5min after epoch boundary before trying to send.
It also checks for CNCLI error message, and if so doesn't mark epoch as sent.