Allow user to run gLiveView from any directory
The current implementation requires that the user run gLiveView.sh from the `$CNODE_HOME/scripts` directory. Meaning that I can't call it directly from whatever directory I am in. For example, by just typing full path, like this: 

`~/> $CNODE_HOME/scripts/gLiveView.sh`

or creating a symlink in my home directory called `~/scripts/` that points to the `$CNODE_HOME/scripts` so I can call directly like this:

`$> ~/scripts/gLiveView.sh`

This patch checks to see if `$CNODE_HOME` is defined and that $CNODE_HOME/scripts is a directory. If so, `PARENT` is set to that directory. Otherwise, it reverts to the previous implementation.