gLiveView 1.27.1 (#1487)
CPU utilization showed an average value since process start. This isn't that useful. This is now replaced by grabbing stats from /proc/stat instead. This is a system metric, not a node metric. But there are no good alternatives to grab just the process cpu utilization. top could be parsed but its a very heavy endpoint to parse. For now I think this is the best option.