[gRest] /datum_info endpoint (#1548)
## Description
<!--- Describe your changes -->
Adding an endpoint for getting datum information based on its hash.
Currently returns only the datum `value` in JSON and `bytes` format. We could add more to it in the future, for now, it seems to be all that's needed.

To consider: make this endpoint accept a list of hashes as bulk input?

## Which issue it fixes?
<!--- Link to issue: Closes #issue-number -->
Closes https://github.com/cardano-community/koios-artifacts/issues/86

## How has this been tested?
<!--- Describe how you tested changes -->
Tested endpoint on guildnet.