Update dockerfile_bin to use ${GUILD_DEPLOY_BRANCH} instead of master (#1730)
## Description
Changes `files/docker/node/dockerfile_bin` to always rely on the
GUILD_DEPLOY_BRANCH for the `guild-deploy.sh` file as well as the
configuration files.

## Where should the reviewer start?


## Motivation and context
Changes merged altering `config-mainnet.json` to
`files/config/mainnet/config.json` leads to failures to build the image.
Can be observed in [this job
run](https://github.com/cardano-community/guild-operators/actions/runs/7690838700).
Changing `dockerfile_bin` to use the **GUILD_DEPLOY_BRANCH** does work.
It also ensures a container that is intended to test a specific set of
changes obtains the **addons**, **scripts**, or **entrypoint.sh** for
the container from the same branch intended for testing.


## Which issue it fixes?
Didn't open an issue, just created the branch and tested the fix. Can
open one if we want the PR to track closing an issue.

## How has this been tested?
The working job from this branch can be observed in [this job
run](https://github.com/cardano-community/guild-operators/actions/runs/7690854231).