CNTools 13.3.4 (#1849)
Bump cardano-hw-cli version requirement to 1.17.0

## Description
cardano-hw-cli has remained in release candidate state with support for
Conway era for a long time. v1.15.0-rc.1 was to be used with Trezor and
v1.16.0-rc.1 for Ledger causing confusion for what version to deploy.
The recently released 1.17.0 version fixes this and is also
auto-deployed with guild-deploy.sh script using `-s w` flag.