Update function: check_cncli_send_tip(), within: healthcheck.sh (#3)
* Update healthcheck.sh

Update function: check_cncli_send_tip()

- Added checks for the tip progressing to reduce false positives when the timeout is reached but the tip hasn't moved.

- Increased the log_entry_timeout to 99, to be just within the container HEATHCHECK timout of 100.

- Improved log output to include current tip, providing more info to the operator.

Co-Authored-By: illuminatus <9167887+TrevorBenson@users.noreply.github.com>

* Update healthcheck.sh

Removed superfluous '-e' from 'echo -e'

* Update healthcheck.sh

renamed uppercase variable to lowercase  to conform to conventions as it's a locally scoped variable.

* Update healthcheck.sh

Added an allowable drift on the tip check as not every tip change is sent to pooltool.

* Update healthcheck.sh

Improved log output to make it a little clearer.

* Update healthcheck.sh

Moved variable 'log_entry_timeout' to be closer to where it's called. It doesn't need to be at the top of the function as it's not intended for the operator to change it.

* Update healthcheck.sh

updated comments

* Update healthcheck.sh

Renamed variable so it better represents what it relates to

* Renamed function

'check_cncli_sendtip' instead of 'check_cncli_send_tip' to align with upstream naming.

* Local variable > user variable

Changed local variable to a user variable. 'log_entry_timeout' > 'NCLI_SENDTIP_LOG_TIMEOUT'

* Local variable > user variable

Changed local variable to a user variable. 'tip_allowed_difference' > 'NCNCLI_SENDTIP_ALLOWED_DIFF'

* Changed container healthcheck params

Reduced container healthcheck interval and timeout to 120s to improve observability.

Co-Authored-By: illuminatus <9167887+TrevorBenson@users.noreply.github.com>

* Renamed variable

Renamed variable to better align with prior naming conventions.

Co-Authored-By: illuminatus <9167887+TrevorBenson@users.noreply.github.com>

* Make use of check_tip()

Make use of check_tip() function to reduce code repetitiveness.

Co-Authored-By: illuminatus <9167887+TrevorBenson@users.noreply.github.com>

* Rename 'TIP' variables

Rename check_tip() variables to make them more generic, allowing the check_tip() function to be used more broadly

Co-Authored-By: illuminatus <9167887+TrevorBenson@users.noreply.github.com>

* Fix function name

Fix incorrect name of function. The function used to be named with an underscore. It was updated but this was missed prior.

* Update variable names

Continuation of updating TIP variable names within check_tip()

Co-Authored-By: illuminatus <9167887+TrevorBenson@users.noreply.github.com>

---------

Co-authored-by: illuminatus <9167887+TrevorBenson@users.noreply.github.com>