Include healthcheck logic for helper scripts running as sidecars (#1842)
## Description
Enhances the healthcheck.sh script to work for checking permissions on
sidecar containers (helper scripts) via the ENTRYPOINT_PROCESS.

## Where should the reviewer start?
1. Copy the updated healthcheck.sh script into the sidecar container in
`/home/guild/.scripts/`.
2. Wait for the 5 minute check interval to occur and confirm if the
sidecar is now shown as healthy.


### Testing different CPU Usage values
* Define a CPU_THRESHOLD environment variable (defaults to `80` %) at a
value you want to mark a container unhealthy when it is exceeded.

### Testing different amount of retries (internal to healthcheck.sh
script).
* Define a RETRIES environment variable (defaults to `20`) at a number
of retries you want to perform if the CPU usage is above the
CPU_THRESHOLD value before exiting non zero

Currently it is a 3 second delay between checks, so 20 retries results
in up to 60 seconds before the healthcheck will exit as unhealthy due to
CPU load.

### Testing different healthcheck values (external to healthcheck.sh
script).
The current HEALTHCHECK of the container image is:
* 5 minutes start period
* 5 minutes interval
* 100 seconds timeout
* 3 retries (default value from being undefined)

Reducing the start period and intervals to something more appropriate
for the sidecar script will result in a much shorter period to determine
the sidecar containers health.

**Make sure to keep the environment variable RETRIES * 3 < container
healthcheck timeout to avoid marking the container unhealthy before the
script will return during periods of high cpu load.**


## Motivation and context
Issue #1841 

## Which issue it fixes?
Closes #1841

## How has this been tested?
1. `docker cp` the script into preview network cncli sync, validate and
leaderlog containers and waiting until the interval runs the script
2. Execute the script with `docker exec` to confirm it reports healthy
3. Monitor the containers until the healthcheck interval occurs and that
they are marked healthy.

## Additional Details

There is a SLEEPING_SCRIPTS array which is used for validate and
leaderlog to still check for the cncli binary, but not consider a sleep
period for validate and leaderlog to be unhealthy. Not 100% sure this is
the best approach, but with sleep periods being variable I felt it was
likely an acceptable middle ground.

Please do not hesitate to suggest an alternative approach to handling
sleeping sidecars healthchecks if you think you have an improvement.

@adamsthws if you could please copy this into your sidecar containers
(and your pool) and report back any results. I am marking this as a
draft PR for the time being until testing is completed, after which if
things look good I will mark it for review and get feedback from others.

Thanks

---------

Co-authored-by: RdLrT <3169068+rdlrt@users.noreply.github.com>
Co-authored-by: Adam Matthews <52178922+adamsthws@users.noreply.github.com>