Ekg cleanup (#1883)
## Description
<!--- Describe your changes -->
With the new tracing infrastructure, we're using PrometheusSimple as
backend without involving cardano-tracer.
If someone has a need for enabling cardano-tracer, they're free to
enable and use it - but our scripts should no longer use/rely on them,
same goes for forwarder and PID listen checks

This also presents an opportunity to perform code cleanup for EKG and
less logic handling.
This PR is expected to follow after #1882 and will have it's delta view
updated once merged

---------

Co-authored-by: Ola Ahlman <ola.ahlman@gmail.com>