Docker stable slim adduser missing (#1891)
## Description
The debian slim image no longer has the adduser wrapper script, causing
the docker build to fail to add the user.

## Which issue it fixes?
Closes #1888 (opened by me, reported to me by @hodlonaut)

---

Additionally this removes the sudo installation and usage during the
build process. The root user can still be used via `docker exec -it -u
root bash` if an operator wishes to manage the container without
rebuilding it.

---------

Co-authored-by: RdLrT <3169068+rdlrt@users.noreply.github.com>